Enum dofus_protocol::game::cosmetic::ObjectSlot

#[repr(i32)]
pub enum ObjectSlot {

    Amulet = 0,
    RingLeft = 1,
    RingRight = 2,
    Belt = 3,
    Boots = 4,
    Petmount = 5,
    Dragoturkey = 6,
    Rhineetle = 7,
    Seemyool = 8,
    Cape = 9,
    Hat = 10,
    Pet = 11,
    Shield = 12,
    Bow = 13,
    Wand = 14,
    Staff = 15,
    Dagger = 16,
    Scythe = 17,
    Axe = 18,
    Lance = 19,
    Hammer = 20,
    Shovel = 21,
    Sword = 22,
    Disguise = 23,
    Wings = 24,
    Shoulders = 25,

}

Variants

Amulet = 0

RingLeft = 1

RingRight = 2

Belt = 3

Boots = 4

Petmount = 5

Dragoturkey = 6

Rhineetle = 7

Seemyool = 8

Cape = 9

Hat = 10

Pet = 11

Shield = 12

Bow = 13

Wand = 14

Staff = 15

Dagger = 16

Scythe = 17

Axe = 18

Lance = 19

Hammer = 20

Shovel = 21

Sword = 22

Disguise = 23

Wings = 24

Shoulders = 25

Implementations

impl ObjectSlot

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ObjectSlot.

pub fn from_i32(value: i32) -> Option<ObjectSlot>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ObjectSlot, or None if value is not a valid variant.

impl ObjectSlot

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.