Enum dofus_protocol::game::game::action::game_action_fight_event::Action

pub enum Action {

    Slide(Slide),
    DodgePointLoss(DodgePointLoss),
    ReflectDamages(ReflectDamages),
    ReduceDamages(ReduceDamages),
    ReflectSpell(ReflectSpell),
    RemovableEffect(RemovableEffect),
    LifePointsLost(LifePointsLost),
    LifePointsGain(LifePointsGain),
    SpellImmunity(SpellImmunity),
    SpellCoolDownVariation(SpellCoolDownVariation),
    Vanish(Vanish),
    Kill(Kill),
    Death(Death),
    TargetedAbility(TargetedAbility),
    Tackled(Tackled),
    PointsVariation(PointsVariation),
    InvisibleDetected(InvisibleDetected),
    TeleportOnSameMap(TeleportOnSameMap),
    ExchangePositions(ExchangePositions),
    SpellRemove(SpellRemove),
    ModifyEffectsDuration(ModifyEffectsDuration),
    StealKama(StealKama),
    ChangeLook(ChangeLook),
    Invisibility(Invisibility),
    Summons(Summons),
    MarkCells(MarkCells),
    UnmarkCells(UnmarkCells),
    TriggerGlyphTrap(TriggerGlyphTrap),
    ActivateGlyphTrap(ActivateGlyphTrap),
    CarryCharacter(CarryCharacter),
    ThrowCharacter(ThrowCharacter),
    DropCharacter(DropCharacter),
    ExecuteScript(ExecuteScript),

}

Variants

Slide(Slide)

DodgePointLoss(DodgePointLoss)

ReflectDamages(ReflectDamages)

ReduceDamages(ReduceDamages)

ReflectSpell(ReflectSpell)

RemovableEffect(RemovableEffect)

LifePointsLost(LifePointsLost)

LifePointsGain(LifePointsGain)

SpellImmunity(SpellImmunity)

SpellCoolDownVariation(SpellCoolDownVariation)

Vanish(Vanish)

Kill(Kill)

Death(Death)

TargetedAbility(TargetedAbility)

Tackled(Tackled)

PointsVariation(PointsVariation)

InvisibleDetected(InvisibleDetected)

TeleportOnSameMap(TeleportOnSameMap)

ExchangePositions(ExchangePositions)

SpellRemove(SpellRemove)

ModifyEffectsDuration(ModifyEffectsDuration)

StealKama(StealKama)

ChangeLook(ChangeLook)

Invisibility(Invisibility)

Summons(Summons)

MarkCells(MarkCells)

UnmarkCells(UnmarkCells)

TriggerGlyphTrap(TriggerGlyphTrap)

ActivateGlyphTrap(ActivateGlyphTrap)

CarryCharacter(CarryCharacter)

ThrowCharacter(ThrowCharacter)

DropCharacter(DropCharacter)

ExecuteScript(ExecuteScript)

Implementations

impl Action

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Action>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.