Enum dofus_protocol::game::party::party_join_error_event::PartyJoinError

#[repr(i32)]
pub enum PartyJoinError {

    Unknown = 0,
    PlayerNotFound = 1,
    PartyNotFound = 2,
    PartyFull = 3,
    PlayerBusy = 4,
    PlayerAlreadyInvited = 5,
    PlayerTooSolicited = 6,
    PlayerLoyal = 7,
    Unmodifiable = 8,
    UnmetCriterion = 9,
    NotEnoughRoom = 10,
    CompositionChanged = 11,
    PlayerInTutorial = 12,

}

Variants

Unknown = 0

PlayerNotFound = 1

PartyNotFound = 2

PartyFull = 3

PlayerBusy = 4

PlayerAlreadyInvited = 5

PlayerTooSolicited = 6

PlayerLoyal = 7

Unmodifiable = 8

UnmetCriterion = 9

NotEnoughRoom = 10

CompositionChanged = 11

PlayerInTutorial = 12

Implementations

impl PartyJoinError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of PartyJoinError.

pub fn from_i32(value: i32) -> Option<PartyJoinError>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a PartyJoinError, or None if value is not a valid variant.

impl PartyJoinError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.